Rename rule consuming preferred write partitioning
"determine preferred writer partitioning" suggested that the rule
determines what is the preferred partitioning. The rule determines
actual partitioning, taking preferred partitioning as the input.

Additionally, "Write" is replaced with "TableWriter" in rule name, to
indicate this works with `TableWriterNode`.
