+++ core/trino-main/src/main/java/io/trino/sql/planner/iterative/rule/ApplyPreferredTableExecutePartitioning.java
@@ -17,35 +17,30 @@ import io.trino.Session;
 import io.trino.matching.Captures;
 import io.trino.matching.Pattern;
 import io.trino.sql.planner.iterative.Rule;
-import io.trino.sql.planner.plan.TableWriterNode;
+import io.trino.sql.planner.plan.TableExecuteNode;
 
 import java.util.Optional;
 
 import static io.trino.SystemSessionProperties.getPreferredWritePartitioningMinNumberOfPartitions;
 import static io.trino.SystemSessionProperties.isUsePreferredWritePartitioning;
 import static io.trino.cost.AggregationStatsRule.getRowsCount;
-import static io.trino.sql.planner.plan.Patterns.tableWriterNode;
+import static io.trino.sql.planner.plan.Patterns.tableExecute;
 import static java.lang.Double.isNaN;
 
 /**
- * Rule verifies if preconditions for using preferred write partitioning are met:
- *  - expected number of partitions to be written (based on table stat) is greater
- *    than or equal to preferred_write_partitioning_min_number_of_partitions session property,
- *  - use_preferred_write_partitioning is set to true.
- *
- * If precondition are met the {@link TableWriterNode} is modified to mark the intention to use preferred write partitioning:
- * value of {@link TableWriterNode#getPreferredPartitioningScheme()} is set as result of {@link TableWriterNode#getPartitioningScheme()}.
+ * Replaces {@link TableExecuteNode} with {@link TableExecuteNode#getPreferredPartitioningScheme()}
+ * with a {@link TableExecuteNode} with {@link TableExecuteNode#getPartitioningScheme()} set.
  */
-public class DeterminePreferredWritePartitioning
-        implements Rule<TableWriterNode>
+public class ApplyPreferredTableExecutePartitioning
+        implements Rule<TableExecuteNode>
 {
-    public static final Pattern<TableWriterNode> WRITER_NODE_WITH_PREFERRED_PARTITIONING = tableWriterNode()
+    public static final Pattern<TableExecuteNode> TABLE_EXECUTE_NODE_WITH_PREFERRED_PARTITIONING = tableExecute()
             .matching(node -> node.getPreferredPartitioningScheme().isPresent());
 
     @Override
-    public Pattern<TableWriterNode> getPattern()
+    public Pattern<TableExecuteNode> getPattern()
     {
-        return WRITER_NODE_WITH_PREFERRED_PARTITIONING;
+        return TABLE_EXECUTE_NODE_WITH_PREFERRED_PARTITIONING;
     }
 
     @Override
@@ -55,7 +50,7 @@ public class DeterminePreferredWritePartitioning
     }
 
     @Override
-    public Result apply(TableWriterNode node, Captures captures, Context context)
+    public Result apply(TableExecuteNode node, Captures captures, Context context)
     {
         int minimumNumberOfPartitions = getPreferredWritePartitioningMinNumberOfPartitions(context.getSession());
         if (minimumNumberOfPartitions <= 1) {
@@ -74,9 +69,9 @@ public class DeterminePreferredWritePartitioning
         return enable(node);
     }
 
-    private Result enable(TableWriterNode node)
+    private static Result enable(TableExecuteNode node)
     {
-        return Result.ofPlanNode(new TableWriterNode(
+        return Result.ofPlanNode(new TableExecuteNode(
                 node.getId(),
                 node.getSource(),
                 node.getTarget(),
@@ -84,10 +79,7 @@ public class DeterminePreferredWritePartitioning
                 node.getFragmentSymbol(),
                 node.getColumns(),
                 node.getColumnNames(),
-                node.getNotNullColumnSymbols(),
                 node.getPreferredPartitioningScheme(),
-                Optional.empty(),
-                node.getStatisticsAggregation(),
-                node.getStatisticsAggregationDescriptor()));
+                Optional.empty()));
     }
 }
